requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-api-python-client==2.116.0
requests==2.31.0
deepgram-sdk==2.12.0
python-dotenv==1.0.1 